refactor: dummyData 파일 이동 및 import 경로 변경
- dummyData 파일을 __dummy-data__로 이동
- 그에 따른 import from 경로 변경
- 그 외 사용하지 않는 import 요소 삭제
