+++ client/src/components/day-transaction-container/DayTransactionContainer.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import DayTransactionCotainer from './DayTransactionContainer';
-import { transactions } from '../../utils/dummyData';
+import { transactions } from '../../__dummy-data__/dummyData';
 
 export default {
   title: 'day-transaction-container/DayTransactionContainer',
