+++ client/src/components/transaction-item/TransactionItem.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import TransactionItem from './TransactionItem';
-import { incomes, expenditures } from '../../utils/dummyData';
+import { incomes, expenditures } from '../../__dummy-data__/dummyData';
 
 export default {
   title: 'transaction-item/transactionItem',
